refactor(RandomNumberGenerator): 패키지 변경
