+++ docs/README.md
@@ -69,18 +69,16 @@
 
 - 생성자의 파라미터로 자동차 리스트를 받아서, 가장 위치 값이 큰 자동차들을 상태 값으로 갖습니다.
 
+### RandomNumberGenerator
+
+- 랜덤 값을 생성하는 객체입니다.
+- 각 인스턴스가 별개의 값을 가질 필요가 없기 때문에 싱글톤으로 구현했습니다.
+
 ## dto
 
 ### CarDto
 
 - 자동차의 상태 값을 view로 출력하기 위해 필요한 DTO 객체입니다.
-
-## util
-
-### RandomNumberGenerator
-
-- 랜덤 값을 생성하는 객체입니다.
-- 각 인스턴스가 별개의 값을 가질 필요가 없기 때문에 싱글톤으로 구현했습니다.
 
 ## view
