+++ src/main/java/racingcar/controller/RaceController.java
@@ -5,7 +5,7 @@ import racingcar.domain.Cars;
 import racingcar.domain.PlayCount;
 import racingcar.domain.Winners;
 import racingcar.dto.CarDto;
-import racingcar.util.RandomNumberGenerator;
+import racingcar.domain.RandomNumberGenerator;
 import racingcar.view.InputViewable;
 import racingcar.view.OutputViewable;
